wayland-egl: allow shipping the library or not
Recently the wayland-egl library and pkg-config file were moved to the
Wayland repository. With that a strange conflict came to be - which one
should be used and when.

The long term goal is to remove the Mesa copies, but with this patch we
allow builders to explicitly select if they want it.

Note: since the header (wayland-egl-backend.h) is now used by C++
people, s/private/driver_private/ was applied.

Cc: Eric Engestrom <eric.engestrom@intel.com>
CC: 18.0 18.1 <mesa-stable@freedesktop.org>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>