wayland-egl: force disable --enable-bundled-wayland-egl
We want to remove all the respective code, so force it off.

As people miss the warning messages produced by autoconf leave the
toggle as-is, thus anyone who explicitly enables it will get an lovely
error and effectively failure ;-)

If anyone is interested in doing the same for meson - patches welcome.

Cc: Eric Engestrom <eric.engestrom@intel.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>