travis: bump required LLVM version to 5.0
We'll bump the number in the build systems shortly. Update the travis
file, first, to avoid intermittent failures.

This effectively removes LLVM 3.9 and 4.0 from the build matrix.

Cc: Juan A. Suarez Romero <jasuarez@igalia.com>
Cc: Dylan Baker <dylan@pnwbakers.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
---
Gents any idea how to resolve the llvm-config workaround we have for
meson? I could not find anything about an override in the meson docs.