meson: bump the minimum LLVM version to 5.0.1
Currently the minimal version is 3.3 with 3.9 for r600+opencl.

We kept it as low, since transitioning to newer ones is pain on Windows.
With issues ranging from build problems, memory leaks and various other
small nitpicks throughout.

Currently we use 5.0.1 for our Appveyor testing, so we can bump the
numbers officially and start ripping out the hacks we had in place.

All recent distribution (versions) have it, with the odd old one left in
the dark. On those platforms, realistically nobody is using mesa-git.

Archlinux: 6.0.1
CentOS: ?
Debian testing: 6.0
Fedora 27: 5.0.1
Gentoo: 5.0.2
OpenSuse 15: 5.0.1
Ubuntu 18.04LTS: 6.0

Cc: Dylan Baker <dylan@pnwbakers.com>
Cc: Eric Engestrom <eric.engestrom@intel.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>